# Hash: optimize `to_a`, `keys` and `values`